+++ src/header/TransferBench.hpp
@@ -3805,7 +3805,6 @@ static bool IsConfiguredGid(union ibv_gid const& gid)
                 transferIdx, i, dstIdx, t.dsts[dstIdx].memRank, expected[i], output[i]};
             }
           }
-          return {ERR_FATAL, "Transfer %d: Unexpected output mismatch for destination %d", transferIdx, dstIdx};
         }
       }
     }